refactor(perf): split /templates + /creation-chronicles into server shell + motion island
37th + 38th in the pattern series.
